ggml-qnn: refine source code structure to make code more clearly
